rolled back 2

+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreStorage.java
@@ -12,7 +12,7 @@ public interface GenreStorage {
 
     Genre getGenreById(Integer id);
 
-    Set<Genre> getFilmGenres(Long filmId);
+    Set<Genre> getFilmGenres(int filmId);
 
     void updateFilmGenres(Film film);
